pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "made-with-gemini"
version = "0.1.0"
description = "This repository contains a collection of tools and examples for building applications with Google Gemini AI."
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Ravi Kumar Chavva", email = "ravikumarchavva@outlook.com" }
]
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-adk>=1.3.0",
    "google-genai>=1.20.0",
    "markdown2>=2.5.3",
    "reportlab>=4.4.1",
    "toolbox-core>=0.2.1",
]

[tool.setuptools]
packages = ["direct_api", "gcp_powered"]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]
